fix: change package name
